POMFSDev_NAS/requirements.txt
Flask
pandas
openpyxl
python-dotenv
apify-client
requests
instagrapi
psycopg2-binary
selenium
beautifulsoup4
lxml
mistralai
undetected-chromedriver
google-cloud-storage
sift-stack-py
markdown
